+++ tools/map_maker/osm_converter/src/OsmJunctionProcessor.cpp
@@ -1,17 +1,8 @@
 // ----------------- BEGIN LICENSE BLOCK ---------------------------------
 //
-// INTEL CONFIDENTIAL
+// Copyright (C) 2017-2020 Intel Corporation
 //
-// Copyright (c) 2017-2019 Intel Corporation
-//
-// This software and the related documents are Intel copyrighted materials, and
-// your use of them is governed by the express license under which they were
-// provided to you (License). Unless the License provides otherwise, you may not
-// use, modify, copy, publish, distribute, disclose or transmit this software or
-// the related documents without Intel's prior written permission.
-//
-// This software and the related documents are provided as is, with no express or
-// implied warranties, other than those that are expressly stated in the License.
+// SPDX-License-Identifier: MIT
 //
 // ----------------- END LICENSE BLOCK -----------------------------------
 
@@ -394,7 +385,7 @@ common::RightOfWay OsmJunctionProcessor::rowForArm(OsmJunctionArm const &arm) co
   {
     if (way.nodes()[i].ref() == arm.mArmId)
     {
-      currentIndex = i;
+      currentIndex = static_cast<int32_t>(i);
     }
   }
 
@@ -429,7 +420,7 @@ common::RightOfWay OsmJunctionProcessor::rowForArm(OsmJunctionArm const &arm) co
       // @todo: change the way to predecessor/successor and continue iteration
       //        for now iterating over one way should be sufficient
     }
-    currentNodeId = way.nodes()[currentIndex].ref();
+    currentNodeId = way.nodes()[static_cast<size_t>(currentIndex)].ref();
   }
 
   return common::RightOfWay::Undefined;
